+++ library/controller/kmp-tor-controller/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/controller/TorController.kt
@@ -27,6 +27,7 @@ import io.matthewnelson.kmp.tor.controller.common.file.Path
 import io.matthewnelson.kmp.tor.controller.common.internal.PlatformUtil
 import io.matthewnelson.kmp.tor.controller.common.internal.isUnixPath
 import io.matthewnelson.kmp.tor.controller.internal.controller.RealTorController
+import io.matthewnelson.kmp.tor.controller.internal.controller.RealTorControlProcessor
 import io.matthewnelson.kmp.tor.controller.internal.getTorControllerDispatcher
 import io.matthewnelson.kmp.tor.controller.internal.toTorController
 import kotlinx.coroutines.withContext
